ML-420: move templated ops...M-O
