+++ libs/ml/include/ml/ops/max_pool.hpp
@@ -17,12 +17,8 @@
 //
 //------------------------------------------------------------------------------
 
-#include "ml/exceptions/exceptions.hpp"
-#include "ml/ops/max_pool_1d.hpp"
-#include "ml/ops/max_pool_2d.hpp"
 #include "ml/ops/ops.hpp"
 
-#include <cassert>
 #include <memory>
 #include <vector>
 
@@ -35,7 +31,7 @@ class MaxPool : public Ops<T>
 {
 public:
   using TensorType    = T;
-  using SizeType      = typename TensorType::SizeType;
+  using SizeType      = fetch::math::SizeType;
   using DataType      = typename TensorType::Type;
   using ArrayPtrType  = std::shared_ptr<TensorType>;
   using VecTensorType = typename Ops<T>::VecTensorType;
@@ -47,35 +43,16 @@ public:
     , stride_size_{stride_size}
   {}
 
-  explicit MaxPool(SPType const &sp)
-    : Ops<T>(sp)
-  {
-    kernel_size_ = sp.kernel_size;
-    stride_size_ = sp.stride_size;
-  }
+  explicit MaxPool(SPType const &sp);
 
   MaxPool(MaxPool const &other) = default;
 
   ~MaxPool() override = default;
 
-  std::shared_ptr<OpsSaveableParams> GetOpSaveableParams() override
-  {
-    SPType sp{};
-    sp.kernel_size = kernel_size_;
-    sp.stride_size = stride_size_;
-    return std::make_shared<SPType>(sp);
-  }
+  std::shared_ptr<OpsSaveableParams> GetOpSaveableParams() override;
 
   std::shared_ptr<fetch::ml::ops::Ops<TensorType>> MakeSharedCopy(
-      std::shared_ptr<fetch::ml::ops::Ops<TensorType>> me) override
-  {
-    FETCH_UNUSED(me);
-    assert(me.get() == this);
-
-    auto copyshare = std::make_shared<MyType>(*this);  // calls default copy constructor of MyType
-
-    return copyshare;
-  }
+      std::shared_ptr<fetch::ml::ops::Ops<TensorType>> me) override;
 
   /**
    * Applies 1D/2D max pooling of kernel_size_ (x kernel_size_) for each channel described here:
@@ -87,13 +64,7 @@ public:
    * number_of_stride_sized_steps_over_input_width)]
    * @return: output tensor parameter
    */
-  void Forward(VecTensorType const &inputs, TensorType &output) override
-  {
-    assert(inputs.size() == 1);
-
-    UpdatePointer(inputs);
-    pool_op_ptr_->Forward(inputs, output);
-  }
+  void Forward(VecTensorType const &inputs, TensorType &output) override;
 
   /**
    * Computes gradient of 2D max pooling of kernel_size_ (x kernel_size) for each channel described
@@ -108,45 +79,9 @@ public:
    * output[0]=input_error[inputs[0].shape]
    */
   std::vector<TensorType> Backward(VecTensorType const &inputs,
-                                   TensorType const &   error_signal) override
-  {
-    assert(inputs.size() == 1);
+                                   TensorType const &   error_signal) override;
 
-    UpdatePointer(inputs);
-    return pool_op_ptr_->Backward(inputs, error_signal);
-  }
-
-  std::vector<SizeType> ComputeOutputShape(VecTensorType const &inputs) const override
-  {
-    assert(inputs.size() == 1);
-    assert(inputs.at(0)->shape().size() == 3 || inputs.at(0)->shape().size() == 4);
-
-    std::vector<SizeType> output_shape;
-
-    // output_shape_[0]=number of output channels
-    output_shape.emplace_back(inputs.at(0)->shape().at(0));
-    // output_shape_[1]=number of stride_size steps over input height
-    output_shape.emplace_back((inputs.at(0)->shape().at(1) - (kernel_size_ - stride_size_)) /
-                              stride_size_);
-
-    // MaxPool1D
-    if (inputs.at(0)->shape().size() == 3)
-    {
-      // output_shape_[2]=batch dimension
-      output_shape.emplace_back(inputs.at(0)->shape().at(2));
-    }
-    // MaxPool2D
-    else
-    {
-      // output_shape_[2]=number of stride_size steps over input width
-      output_shape.emplace_back((inputs.at(0)->shape().at(2) - (kernel_size_ - stride_size_)) /
-                                stride_size_);
-      // output_shape_[3]=batch dimension
-      output_shape.emplace_back(inputs.at(0)->shape().at(3));
-    }
-
-    return output_shape;
-  }
+  std::vector<SizeType> ComputeOutputShape(VecTensorType const &inputs) const override;
 
   static constexpr OpType OpCode()
   {
@@ -162,46 +97,6 @@ private:
 
   void UpdatePointer(VecTensorType const &inputs);
 };
-
-/**
- * Update pointer to pooling op depending on input size
- * @tparam TensorType
- * @param inputs
- */
-template <typename TensorType>
-void MaxPool<TensorType>::UpdatePointer(VecTensorType const &inputs)
-{
-  switch (inputs.at(0)->shape().size())
-  {
-    // MaxPool1D
-  case static_cast<SizeType>(3):
-  {
-    if (!pool_op_ptr_ || pool_2d_)
-    {
-      pool_op_ptr_ =
-          std::make_shared<fetch::ml::ops::MaxPool1D<TensorType>>(kernel_size_, stride_size_);
-      pool_2d_ = false;
-    }
-    break;
-  }
-    // MaxPool2D
-  case static_cast<SizeType>(4):
-  {
-    if (!pool_op_ptr_ || !pool_2d_)
-    {
-      pool_op_ptr_ =
-          std::make_shared<fetch::ml::ops::MaxPool2D<TensorType>>(kernel_size_, stride_size_);
-      pool_2d_ = true;
-    }
-    break;
-  }
-
-  default:
-  {
-    throw fetch::ml::exceptions::InvalidMode("Unsupported data shape");
-  }
-  }
-}
 
 }  // namespace ops
 }  // namespace ml
