+++ libs/ml/include/ml/ops/max_pool_2d.hpp
@@ -19,7 +19,7 @@
 
 #include "ml/ops/ops.hpp"
 
-#include <cassert>
+#include <memory>
 #include <vector>
 
 namespace fetch {
@@ -43,33 +43,14 @@ public:
     , stride_size_{stride_size}
   {}
 
-  explicit MaxPool2D(SPType const &sp)
-    : Ops<T>(sp)
-  {
-    kernel_size_ = sp.kernel_size;
-    stride_size_ = sp.stride_size;
-  }
+  explicit MaxPool2D(SPType const &sp);
 
   ~MaxPool2D() override = default;
 
-  std::shared_ptr<OpsSaveableParams> GetOpSaveableParams() override
-  {
-    SPType sp{};
-    sp.kernel_size = kernel_size_;
-    sp.stride_size = stride_size_;
-    return std::make_shared<SPType>(sp);
-  }
+  std::shared_ptr<OpsSaveableParams> GetOpSaveableParams() override;
 
   std::shared_ptr<fetch::ml::ops::Ops<TensorType>> MakeSharedCopy(
-      std::shared_ptr<fetch::ml::ops::Ops<TensorType>> me) override
-  {
-    FETCH_UNUSED(me);
-    assert(me.get() == this);
-
-    auto copyshare = std::make_shared<MyType>(*this);  // calls default copy constructor of MyType
-
-    return copyshare;
-  }
+      std::shared_ptr<fetch::ml::ops::Ops<TensorType>> me) override;
   /**
    * Applies 2D max pooling of kernel_size_ x kernel_size_ for each channel described here:
    * http://ais.uni-bonn.de/papers/icann2010_maxpool.pdf
@@ -79,54 +60,7 @@ public:
    * number_of_stride_sized_steps_over_input_height x number_of_stride_sized_steps_over_input_width]
    * @return: output tensor parameter
    */
-  void Forward(VecTensorType const &inputs, TensorType &output) override
-  {
-    assert(inputs.size() == 1);
-    // Input must be a 4D tensor [C x W x H x N]
-    assert(inputs.at(0)->shape().size() == 4);
-    assert(output.shape() == ComputeOutputShape(inputs));
-
-    SizeType iterw;
-    SizeType iterh;
-    DataType val;
-    DataType max;
-    auto     oit = output.begin();
-
-    for (SizeType n_i{0}; n_i < output.shape().at(3); n_i++)  // iterate over batch
-    {
-      for (SizeType ih{0}; ih < output.shape().at(2); ih++)  // Iterate height over kernel stride
-      {
-        iterh = ih * stride_size_;
-
-        for (SizeType iw{0}; iw < output.shape().at(1); iw++)  // Iterate width over kernel stride
-        {
-          iterw = iw * stride_size_;
-
-          for (SizeType c{0}; c < output.shape().at(0); ++c)  // Iterate over output channels
-          {
-            max = inputs.at(0)->At(c, iterw, iterh, n_i);
-
-            // Get maximum value on kernel_size_ x kernel_size_ window
-            for (SizeType jw{0}; jw < kernel_size_; jw++)  // Iterate over kernel width
-            {
-              for (SizeType jh{0}; jh < kernel_size_; jh++)  // Iterate over kernel width
-              {
-                val = inputs.at(0)->At(c, iterw + jw, iterh + jh, n_i);
-                if (val > max)
-                {
-                  max = val;
-                }
-              }
-            }
-
-            // Set maximum value for each [kernel_size_ x kernel_size_] window to output
-            *oit = max;
-            ++oit;
-          }
-        }
-      }
-    }
-  }
+  void Forward(VecTensorType const &inputs, TensorType &output) override;
 
   /**
    * Computes gradient of 2D max pooling of kernel_size_ x kernel_size for each channel described
@@ -140,80 +74,9 @@ public:
    * output[0]=input_error[inputs[0].shape]
    */
   std::vector<TensorType> Backward(VecTensorType const &inputs,
-                                   TensorType const &   error_signal) override
-  {
-    assert(inputs.size() == 1);
-    assert(error_signal.shape() == ComputeOutputShape(inputs));
-    TensorType return_signal{inputs.at(0)->shape()};
+                                   TensorType const &   error_signal) override;
 
-    SizeType iterh;
-    SizeType iterw;
-    DataType max;
-    DataType val;
-    SizeType max_iterw;
-    SizeType max_iterh;
-    auto     erit = error_signal.cbegin();
-    for (SizeType n_i{0}; n_i < error_signal.shape().at(2); n_i++)  // iterate over batch
-    {
-      // Iterate width over kernel stride
-      for (SizeType iw{0}; iw < error_signal.shape().at(1); iw++)
-      {
-        iterw = iw * stride_size_;
-        // Iterate height over kernel stride
-        for (SizeType ih{0}; ih < error_signal.shape().at(2); ih++)
-        {
-          iterh = ih * stride_size_;
-          // Iterate over output channels
-          for (SizeType c{0}; c < error_signal.shape().at(0); ++c)
-          {
-            max       = inputs.at(0)->At(c, iterw, iterh, n_i);
-            max_iterw = iterw;
-            max_iterh = iterh;
-
-            // Find max node
-            for (SizeType jw{0}; jw < kernel_size_; jw++)  // Iterate over kernel width
-            {
-              for (SizeType jh{0}; jh < kernel_size_; jh++)  // Iterate over kernel width
-              {
-
-                val = inputs.at(0)->At(c, iterw + jw, iterh + jh, n_i);
-                if (val > max)
-                {
-                  max       = val;
-                  max_iterw = iterw + jw;
-                  max_iterh = iterh + jh;
-                }
-              }
-            }
-
-            // Add error to max node
-            return_signal(c, max_iterw, max_iterh, n_i) =
-                return_signal(c, max_iterw, max_iterh, n_i) + *erit;
-            ++erit;
-          }
-        }
-      }
-    }
-
-    return {return_signal};
-  }
-
-  std::vector<SizeType> ComputeOutputShape(VecTensorType const &inputs) const override
-  {
-    std::vector<SizeType> output_shape;
-
-    // output_shape_[0]=number of output channels
-    output_shape.emplace_back(inputs.at(0)->shape().at(0));
-    // output_shape_[1]=number of stride_size steps over input height
-    output_shape.emplace_back((inputs.at(0)->shape().at(1) - (kernel_size_ - stride_size_)) /
-                              stride_size_);
-    // output_shape_[2]=number of stride_size steps over input width
-    output_shape.emplace_back((inputs.at(0)->shape().at(2) - (kernel_size_ - stride_size_)) /
-                              stride_size_);
-    // output_shape_[3]=batch dimension
-    output_shape.emplace_back(inputs.at(0)->shape().at(3));
-    return output_shape;
-  }
+  std::vector<SizeType> ComputeOutputShape(VecTensorType const &inputs) const override;
 
   static constexpr OpType OpCode()
   {
