+++ libs/ml/include/ml/ops/one_hot.hpp
@@ -17,10 +17,8 @@
 //
 //------------------------------------------------------------------------------
 
-#include "math/one_hot.hpp"
 #include "ml/ops/ops.hpp"
 
-#include <cassert>
 #include <utility>
 #include <vector>
 
@@ -33,7 +31,7 @@ class OneHot : public fetch::ml::ops::Ops<T>
 {
 public:
   using TensorType    = T;
-  using SizeType      = typename TensorType::SizeType;
+  using SizeType      = fetch::math::SizeType;
   using DataType      = typename TensorType::Type;
   using ArrayPtrType  = std::shared_ptr<TensorType>;
   using VecTensorType = typename Ops<T>::VecTensorType;
@@ -55,74 +53,20 @@ public:
     , off_value_(off_value)
   {}
 
-  explicit OneHot(SPType const &sp)
-    : Ops<T>(sp)
-  {
-    depth_     = sp.depth;
-    axis_      = sp.axis;
-    on_value_  = sp.on_value;
-    off_value_ = sp.off_value;
-  }
+  explicit OneHot(SPType const &sp);
 
   ~OneHot() override = default;
 
-  std::shared_ptr<OpsSaveableParams> GetOpSaveableParams() override
-  {
-    SPType sp{};
-    sp.depth     = depth_;
-    sp.axis      = axis_;
-    sp.on_value  = on_value_;
-    sp.off_value = off_value_;
-
-    return std::make_shared<SPType>(sp);
-  }
+  std::shared_ptr<OpsSaveableParams> GetOpSaveableParams() override;
 
   std::shared_ptr<fetch::ml::ops::Ops<TensorType>> MakeSharedCopy(
-      std::shared_ptr<fetch::ml::ops::Ops<TensorType>> me) override
-  {
-    FETCH_UNUSED(me);
-    assert(me.get() == this);
-
-    auto copyshare = std::make_shared<MyType>(*this);  // calls default copy constructor of MyType
-
-    return copyshare;
-  }
-  void Forward(VecTensorType const &inputs, TensorType &output) override
-  {
-    assert(inputs.size() == 1);
-    assert(output.shape() == this->ComputeOutputShape(inputs));
-
-    fetch::math::OneHot(output, *(inputs.at(0)), depth_, axis_, on_value_, off_value_);
-  }
+      std::shared_ptr<fetch::ml::ops::Ops<TensorType>> me) override;
+  void Forward(VecTensorType const &inputs, TensorType &output) override;
 
   std::vector<TensorType> Backward(VecTensorType const &inputs,
-                                   TensorType const &   error_signal) override
-  {
-    FETCH_UNUSED(error_signal);
-    assert(inputs.size() == 1);
-    assert(error_signal.shape() == this->ComputeOutputShape(inputs));
+                                   TensorType const &   error_signal) override;
 
-    // No derivative defined for OneHotOp
-    return {TensorType{inputs.at(0)->shape()}};
-  }
-
-  std::vector<SizeType> ComputeOutputShape(VecTensorType const &inputs) const override
-  {
-    assert(inputs.size() == 1);
-
-    std::vector<SizeType> shape = inputs.at(0)->shape();
-
-    if (axis_ == shape.size())
-    {
-      shape.emplace_back(depth_);
-    }
-    else
-    {
-      shape.insert(shape.begin() + static_cast<math::PtrDiffType>(axis_), depth_);
-    }
-
-    return shape;
-  }
+  std::vector<SizeType> ComputeOutputShape(VecTensorType const &inputs) const override;
 
   static constexpr OpType OpCode()
   {
